+++ packages/lib/src/components/Dropin/components/PaymentMethod/PaymentMethodItem/PaymentMethodItem.tsx
@@ -128,6 +128,7 @@ class PaymentMethodItem extends Component<Readonly<PaymentMethodItemProps>> {
                             inline
                             variant="link"
                             onClick={this.toggleDisableConfirmation}
+                            onKeyPress={stopPropagationForActionKeys}
                             onKeyDown={stopPropagationForActionKeys}
                             ariaExpanded={this.state.showDisableStoredPaymentMethodConfirmation}
                             ariaControls={disableConfirmationId}