+++ packages/lib/src/components/internal/BaseElement/BaseElement.ts
@@ -1,4 +1,4 @@
-import { ComponentChild, h, render, TargetedKeyboardEvent } from 'preact';
+import { ComponentChild, h, render } from 'preact';
 import uuid from '../../../utils/uuid';
 import AdyenCheckoutError from '../../../core/Errors/AdyenCheckoutError';
 import { NO_CHECKOUT_ATTEMPT_ID } from '../../../core/Analytics/constants';
@@ -47,7 +47,7 @@ abstract class BaseElement<P extends BaseElementProps> implements IBaseElement {
         this.core = checkout;
         this.buildElementProps(props);
 
-        this.handleKeyDown = this.handleKeyDown.bind(this);
+        this.handleKeyPress = this.handleKeyPress.bind(this);
     }
 
     protected buildElementProps(componentProps?: P) {
@@ -70,11 +70,13 @@ abstract class BaseElement<P extends BaseElementProps> implements IBaseElement {
         return {};
     }
 
-    protected submitAnalytics(_analyticsObj?: AbstractAnalyticsEvent) {
+    /* eslint-disable-next-line */
+    protected submitAnalytics(analyticsObj?: AbstractAnalyticsEvent) {
         return null;
     }
 
-    protected handleKeyDown(_e: TargetedKeyboardEvent<HTMLInputElement>) {
+    /* eslint-disable-next-line */
+    protected handleKeyPress(e: h.JSX.TargetedKeyboardEvent<HTMLInputElement>) {
         return null;
     }
 
@@ -143,7 +145,7 @@ abstract class BaseElement<P extends BaseElementProps> implements IBaseElement {
         this._node = node;
 
         // Add listener for key press events, notably 'Enter' key presses
-        on(this._node, 'keydown', this.handleKeyDown, false);
+        on(this._node, 'keypress', this.handleKeyPress, false);
 
         this._component = this.render();
 
@@ -170,7 +172,7 @@ abstract class BaseElement<P extends BaseElementProps> implements IBaseElement {
      */
     public unmount(): this {
         // Remove listener
-        off(this._node, 'keydown', this.handleKeyDown);
+        off(this._node, 'keypress', this.handleKeyPress);
 
         if (this._node) {
             render(null, this._node);